refactor: rename start_screen
