+++ lib/games/tugak_catching/views/start_screen.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import '../../shared/widgets/start_screen.dart';
+import '../../shared/widgets/base_start_screen.dart';
 import './game_screen.dart';
 
 class TugakStartScreen extends StatelessWidget {
